pyproject.toml
[project]
name = "quishing-with-ml"
version = "0.1.0"
description = "QR Code Phishing Detection using Deep Learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "wandb>=0.15.0",
    "tqdm>=4.65.0",
]
